flesh out commands: init-*, build-*, test-*
Note: basically, bin/test-service is working, which

  - copies the source files and lib/.cache to build, and generates env var files
    (bin/build-service)

  - lints the copied files;

  - copies the test image to build, and builds and runs it

Note: local packages are built into wheels and the wheels are then available
to the requirements files. This means they can be properly specified with
version ranges etc., not as simple paths in the requirements files.

Note: the init-* calls related to GCP have not been tested yet.